requirements.txt
absl-py==0.7.1
apturl==0.5.2
astor==0.7.1
audioread==2.1.9
backcall==0.1.0
beautifulsoup4==4.4.1
bleach==3.0.2
blessings==1.7
blinker==1.3
block.bootstrap.pytorch==0.1.6
bootstrap.pytorch==0.0.13
boto3==1.9.83
botocore==1.12.83
Brlapi==0.6.4
bsddb3==6.1.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
checkbox-support==0.22
click==7.1.2
command-not-found==0.3
cryptography==1.2.3
cycler==0.10.0
Cython==0.29.13
decorator==4.3.0
defer==1.0.6
defusedxml==0.5.0
devscripts===2.16.2ubuntu3
dill==0.3.3
docutils==0.14
efficientnet-pytorch==0.6.3
entrypoints==0.2.3
facenet-pytorch==2.5.1
feedparser==5.1.3
fire==0.3.1
flake8==2.5.4
future==0.18.2
gast==0.2.2
google-pasta==0.1.4
gpustat==0.5.0
graphviz==0.8.4
grpcio==1.19.0
guacamole==0.9.2
h5py==2.9.0
html5lib==0.999
httplib2==0.9.1
idna==2.10
imageio==2.3.0
importlib-metadata==2.1.1
inflect==3.0.2
ipdb==0.13.4
ipykernel==5.1.0
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.8
jmespath==0.9.3
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
language-selector==0.1
librosa==0.7.2
llvmlite==0.29.0
louis==2.6.4
lxml==3.5.0
Magic-file-extensions==0.2
Mako==1.0.3
Markdown==3.1
MarkupSafe==0.23
matplotlib==3.0.2
mccabe==0.2.1
mistune==0.8.4
mock==2.0.0
multiprocess==0.70.11.1
munch==2.5.0
mutagen==1.45.1
mxnet-cu90==1.6.0.post0
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.4
nltk==3.3
notebook==5.7.0
numba==0.45.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
oauthlib==1.0.3
onboard==1.2.0
opencv-contrib-python==4.2.0.34
opencv-python==3.4.3.18
openslide-python==1.1.0
padme==1.1.1
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pbr==5.1.3
pep8==1.7.0
pexpect==4.0.1
pickleshare==0.7.5
Pillow==5.2.0
plainbox==0.25
plotly==4.12.0
pretrainedmodels==0.7.4
progress==1.5
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.13.0
psutil==5.4.7
ptyprocess==0.5
py-rouge==1.1
pyasn1==0.1.9
pycparser==2.20
pycups==1.9.73
pycurl==7.43.0
pyflakes==1.1.0
Pygments==2.2.0
pygobject==3.20.0
PyJWT==1.3.0
PyOpenGL==3.0.2
pyparsing==2.0.3
python-apt==1.1.0b1+ubuntu0.16.4.9
python-dateutil==2.7.5
python-debian==0.1.27
python-speech-features==0.6
python-systemd==231
pytorch-pretrained-bert==0.4.0
pytorch-ssim==0.1
pytz==2018.7
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==17.1.2
qtconsole==4.4.2
reportlab==3.3.0
requests==2.25.1
resampy==0.2.2
retrying==1.3.3
rod-align-api==0.0.0
roi-align-api==0.0.0
s3transfer==0.1.13
scenedetect==0.5.1
scikit-image==0.15.0
scikit-learn==0.20.0
scipy==1.4.1
screen-resolution-extra==0.0.0
seaborn==0.9.1
Send2Trash==1.5.0
sessioninstaller==0.0.0
six==1.11.0
skipthoughts==0.0.1
sklearn==0.0
SoundFile==0.10.3.post1
ssh-import-id==5.5
system-service==0.3
tabulate==0.8.2
tb-nightly==1.14.0a20190301
tensorboard==1.13.1
tensorboardX==2.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.3.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
textblob==0.15.3
tf==1.0.0
tf-estimator-nightly==1.14.0.dev2019030115
torch==1.4.0
torchaudio==0.4.0
torchlight==0.0.1
torchsummary==1.5.1
torchvision==0.5.0
torchviz==0.0.1
tornado==5.1.1
tqdm==4.28.1
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.35
Unidecode==1.1.1
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-gdrive==0.7
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.24.1
usb-creator==0.3.0
virtualenv==15.0.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.1
widgetsnbextension==3.4.2
wordcloud==1.7.0
xdiagnose==3.8.4.1
xkit==0.0.0
XlsxWriter==0.7.3
zipp==1.2.0
